requirements.txt
beautifulsoup4==4.9.3
certifi==2021.10.8
chardet==4.0.0
google==3.0.0
html5lib==1.1
idna==2.10
pafy==0.5.5
pypng==0.0.21
PyQRCode==1.2.1
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
pyshorteners==1.0.1
pytube==11.0.1
pyzbar==0.1.8
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
texttable==1.6.4
tk==0.1.0
urllib3==1.26.7
webencodings==0.5.1
youtube-dl==2021.6.6
